+++ CHANGELOG.md
@@ -8,6 +8,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 
+## [0.1.1] - 2023-11-30
+Feature release.
+
+### Changed
+- Saving `DataFrameDescriber` as Excel no longer require the file extension.
+- Fail when trying to clobber `DataFrameDescriber` metadata and columns.
+
 ### Added
 - Methods to "re-hydrate" `DataDescriber` and `DataFrameDescriber` instances
   previously dumped to the file system.
@@ -15,6 +22,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   using `Table`.
 - CLI feature to write formatted tables using YAML and CSV files as an Excel
   file.
+- Feature to save `DataFrameDescriber` as Excel from the command line.
+
 
 ## [0.1.0] - 2023-08-16
 Downstream moderate risk update release.
@@ -34,7 +43,8 @@ Downstream moderate risk update release.
 
 
 <!-- links -->
-[Unreleased]: https://github.com/plandes/datdesc/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/plandes/datdesc/compare/v0.1.1...HEAD
+[0.1.1]: https://github.com/plandes/datdesc/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/plandes/datdesc/compare/v0.0.3...v0.1.0
 [0.0.3]: https://github.com/plandes/datdesc/compare/v0.0.2...v0.0.3
 [0.0.2]: https://github.com/plandes/datdesc/compare/v0.0.1...v0.0.2